requirements.txt
streamlit
pandas
plotly
requests
fastapi
uvicorn
openai
python-dotenv